[EDP] Using trust delegation for Swift authentication
For access to Swift objects, Sahara should be changed to generate a
Keystone trust between the user with access to those objects and the
Sahara admin user. The trust would be established based on the user’s
membership in the project that contains the Swift objects. Using this
trust the Sahara user could generate authentication tokens to access the
Swift objects. When access is no longer needed the trust can be revoked
thus invalidating the tokens.

blueprint edp-swift-trust-authentication

Change-Id: Ia48cd573f4c3ca9a7cbe1edb652f7a27b03615f9