[EDP] Add options supporting DataSource identifiers in job_configs
In some cases it would be convenient if users had a way to pass a DataSource
object as a job configuration value to take advantage of the data codified in
the object instead of copying and specifying that information manually.  This
specification describes options that allow users to reference DataSource
objects in configuration values, parameters, and arguments in a JobExecution
record.

blueprint edp-data-sources-in-job-configs

Change-Id: Ib07e667e019c0002fe11ac1889751844e7435bd6