Unified Map to Define Job Interface
This specification proposes the addition of an "interface" map to the API for
job creation, such that the operator registering a job can define a unified,
human-readable way to pass in all arguments, parameters, and configurations
that the execution of that job may require or accept. This will allow
platform-agnostic wizarding at the job execution phase and allows users to
document use of their own jobs once in a persistent, standardized format.

Change-Id: I49f4c9ed80158522bf41ba224d2c7cc4c54011cc
Describes: blueprint unified-job-interface-map