Enable Spark jobs to access Swift URLs
Spark uses Hadoop filesystem libraries to dereference input and output URLs.
Consequently, Spark can access Swift filesystem URLs if the Hadoop Swift JAR
is included in the image and a Spark job’s Hadoop configuration includes the
necessary Swift credentials. This specification describes a method of
transparently adding Swift credentials to the Hadoop configuration of a Spark
job so that the job source code does not need to be altered and recompiled to
access Swift URLs.

Partially-Implements: blueprint edp-spark-swift-integration
Change-Id: I019e2b841997de0444ef7222172baa57b7bdb45f