Adding improved secret storage spec
There are several secrets (for example passwords) that Sahara uses with
respect to deployed frameworks which are currently stored in its
database. This specification proposes the creation of a Barbican integration
module that will allow Sahara to offload secret storage to that service.
The initial focus of this offloading effort will focus on the storage of
Swift credentials.

SecurityImpact

Change-Id: I7b5b38b5b10f7e88893dca8792a298537b69cf05
Partial-implements: bp improved-secret-storage