Add data source subtypes to unified job interface map spec
To allow further validation of input and output data sources,
two new types are being added to the value_type enumeration in
the unified job interface map spec. Note that the "data_source"
type is also being retained to allow path validation to be
bypassed if desired.

Describes: blueprint unified-job-interface-map

Change-Id: If1dca07bab03838c24e14f6a25a0cec9b8fafcad