Manila as a binary store
Network and distributed filesystems are a useful means of sharing files among
a distributed architecture. This core use case makes them an excellent
candidate for storage of job binaries.

While internal database storage and Swift storage are sensible options for
binary storage, the addition of Manila integration allows it as a third option
for job binary storage and retrieval. This specification details how this
will be implemented.

Change-Id: Ib005d73ba2eefd55cd79feaf4e796d9a98841926
Describes: blueprint manila-as-binary-store