Manila as a runtime data source
Using manila nfs shares and the mounting features developed for use with
job binaries, it should be feasible to use nfs shares to host input
and output data. This will allow data to be referenced through local
filesystem paths as a another simple alternative to hdfs or swift storage.

Change-Id: I374cd3cb47f4b80537a41a5d11402a65a399914b
Partial-implements: bp manila-as-a-data-source