Modify 'is_default' behavior relative to 'is_protected' for templates
With the addition of the general *is_protected* field on all objects,
the *is_default* field on cluster and node group templates is semi-redundant.
Its semantics should be changed; specifically, gating of update and
delete operations should be handled exclusively by *is_protected*. The
*is_default* field should still be used to identify default templates for bulk
update and delete operations by the sahara-templates tool.

Partial-implements: blueprint replace-is-default
Change-Id: I9808df9546636d755fc9c999ebee45e1ec9f2b1a