SPI Method to Validate Images
This specification details the addition of a method to the Sahara plugin SPI
to validate that a chosen image is up to the specification that the plugin
requires. While it is not expected that plugin writers will be able to test
the image deeply enough to ensure that a given arbitrary image will succeed
in cluster generation and be functional in all contexts, it is hoped that by
implementing this method well, plugin authors can provide a well-defined,
machine-actionable contract which will be versioned with the plugin itself.

Describes: blueprint validate-image-spi

Change-Id: I9dbc40cde6be077162e7a621866b567783cafecc