Spec to reorganize the UI into tabs
This spec is to convert the 10 panel approach in
the sahara UI down to 2 panels that will each contain
several tabs.  One tab for each of the current panels.
It represents a significant reorganization of the layout
of the UI and should be a good step forward in terms
of usability.

Change-Id: Ia463b26c5ddf39dcc1cba3da54b7bee479953e0b