Decommission of specific node
When facing issues with a cluster it can be useful to remove an specific
instance. The way that Sahara is constructed today allows the user to scale a
cluster down but it will choose a random instance from the selected node group
to be removed. We want to give the users the opportunity to choose which
instance(s) he/she would like to remove from the cluster.

Change-Id: I885f0583b97c29abb71888114597a7da8deaf4b3